refactor: インフラ層の構成を再編
