+++ Assets/YukimaruGames/Terminal/Runtime/Infrastructure/Accessors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 321603a8adc9bd5478554680d75341b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
