+++ Assets/YukimaruGames/Terminal/Runtime/Infrastructure/Accessors/ColorPaletteAccessor.cs.meta
