+++ Assets/YukimaruGames/Terminal/Runtime/Infrastructure/Discovery.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52034dc640f7d0d41be04a886856be0c
+guid: 8e8c4c9515bfaf140b833defa3ae0e8f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
