+++ Assets/YukimaruGames/Terminal/Runtime/Infrastructure/Factories.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 212e51ced3035fa4ab5b8ba56056b4d5
+guid: 309995dd28c4db54cbf3c18aee9841f1
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
