+++ Assets/YukimaruGames/Terminal/Runtime/Infrastructure/Repositories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8df115e3768aad14ea00b6379f4e183d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
