+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Installer/TerminalStandardInstaller.cs
@@ -19,8 +19,10 @@ using YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Repositories;
 using YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services;
 using YukimaruGames.Terminal.Domain.Repositories;
 using YukimaruGames.Terminal.Domain.Services;
-using YukimaruGames.Terminal.Infrastructure.Commands;
-using YukimaruGames.Terminal.Infrastructure.UI;
+using YukimaruGames.Terminal.Infrastructure.Accessors;
+using YukimaruGames.Terminal.Infrastructure.Discoverer;
+using YukimaruGames.Terminal.Infrastructure.Factories;
+using YukimaruGames.Terminal.Infrastructure.Repositories;
 using YukimaruGames.Terminal.Presentation.Accessors;
 using YukimaruGames.Terminal.Presentation.Animators;
 using YukimaruGames.Terminal.Presentation.Constants;
